tarfs: follow hardlinks in `ReadFile`
In `ReadFile`, we either follow symlinks or attempt to read the
amount of data from the tar as specified by the Size field in
`tar.Header`. When we come across a hardlink, however, we
essentially tell the caller that the file is empty since the Size
is 0.

These changes to ReadFile are largely based on the changes
@hdonnay made to Open for handling hardlinks: 48317f8

Signed-off-by: Brad Lugo <blugo@redhat.com>